% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{dyRup_palettes}
\alias{dyRup_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 7.
}
\usage{
dyRup_palettes
}
\description{
Complete list of palettes
}
\keyword{datasets}
