% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{dyRup}
\alias{dyRup}
\title{A Wes Anderson palette generator}
\usage{
dyRup(palette, n, type = c("discrete", "continuous"))
}
\arguments{
\item{palette}{Name of desired palette. Choices are:
\code{devoid_of_all_desires}, \code{calm_down_woman},  \code{warm_modern},
\code{brighteyes}, \code{dark_musk},  \code{braveheart}, \code{soft_n_hazy}}

\item{n}{Number of colors desired. Almost all palettes have 7 colors.
But hopefully we'll add more palettes soon.
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes from Wes Anderson movies.
}
